+++ src/Mpociot/BotMan/Drivers/FacebookDriver.php
@@ -2,6 +2,7 @@
 
 namespace Mpociot\BotMan\Drivers;
 
+use Mpociot\BotMan\User;
 use Mpociot\BotMan\Answer;
 use Mpociot\BotMan\Message;
 use Mpociot\BotMan\Question;
@@ -220,5 +221,15 @@ class FacebookDriver extends Driver
     public function isConfigured()
     {
         return ! is_null($this->config->get('facebook_token'));
+    }
+
+    /**
+     * Retrieve User information
+     * @param Message $matchingMessage
+     * @return User
+     */
+    public function getUser(Message $matchingMessage)
+    {
+        return new User($matchingMessage->getChannel());
     }
 }