+++ tests/Drivers/FacebookDriverTest.php
@@ -73,6 +73,21 @@ class FacebookDriverTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
+    public function it_returns_the_user_object()
+    {
+        $request = '{"object":"page","entry":[{"id":"111899832631525","time":1480279487271,"messaging":[{"sender":{"id":"1433960459967306"},"recipient":{"id":"111899832631525"},"timestamp":1480279487147,"message":{"mid":"mid.1480279487147:4388d3b344","seq":36,"text":"Hi Julia"}}]}]}';
+        $driver = $this->getDriver($request);
+
+        $message = $driver->getMessages()[0];
+        $user = $driver->getUser($message);
+        
+        $this->assertSame($user->getId(), '1433960459967306');
+        $this->assertNull($user->getFirstName());
+        $this->assertNull($user->getLastName());
+        $this->assertNull($user->getUsername());
+    }
+
+    /** @test */
     public function it_returns_an_empty_message_if_nothing_matches()
     {
         $request = '';